+++ database/README.md
@@ -1,0 +1,10 @@
+# Database Documentation
+This folder contains the scripts and resources for setting up the database for the court transcript data pipeline.
+
+## 📚 Folder Contents
+
+- `judges_seed.py`: Script to seed the database with judge names data.
+
+- `schema.sql`: SQL schema for the database.
+
+## 🛠️ Database Setup Instructions
