+++ database/schema.sql
@@ -13,7 +13,7 @@ DROP TABLE IF EXISTS tag;
 
 CREATE TABLE tag (
     tag_id INT GENERATED ALWAYS AS IDENTITY,
-    tag_name VARCHAR(50) UNIQUE NOT NULL,
+    tag_name VARCHAR(250) UNIQUE NOT NULL,
     PRIMARY KEY (tag_id)
 );
 
@@ -31,7 +31,7 @@ CREATE TABLE court (
 
 CREATE TABLE participant (
     participant_id INT GENERATED ALWAYS AS IDENTITY,
-    participant_name VARCHAR(150) UNIQUE NOT NULL,
+    participant_name VARCHAR(512) UNIQUE,
     PRIMARY KEY (participant_id)
 );
 
@@ -43,13 +43,13 @@ CREATE TABLE verdict (
 
 CREATE TABLE law_firm (
     law_firm_id INT GENERATED ALWAYS AS IDENTITY,
-    law_firm_name VARCHAR(50) UNIQUE NOT NULL,
+    law_firm_name VARCHAR(255) UNIQUE,
     PRIMARY KEY (law_firm_id)
 );
 
 CREATE TABLE lawyer (
     lawyer_id INT GENERATED ALWAYS AS IDENTITY,
-    lawyer_name VARCHAR(100),
+    lawyer_name VARCHAR(200),
     law_firm_id INT,
     UNIQUE(lawyer_name, law_firm_id),
     PRIMARY KEY (lawyer_id),
@@ -57,13 +57,13 @@ CREATE TABLE lawyer (
 );
 
 CREATE TABLE court_case (
-    court_case_id VARCHAR(50) UNIQUE NOT NULL,
+    court_case_id VARCHAR(250) UNIQUE NOT NULL,
     summary TEXT,
     verdict_id INT,
-    title VARCHAR(150),
+    title VARCHAR(512),
     court_date DATE,
-    case_number VARCHAR(50) UNIQUE NOT NULL,
-    case_url VARCHAR(50),
+    case_number VARCHAR(250),
+    case_url VARCHAR(512),
     court_id INT,
     verdict_summary TEXT,
     PRIMARY KEY (court_case_id),
@@ -72,7 +72,7 @@ CREATE TABLE court_case (
 );
 
 CREATE TABLE tag_assignment (
-    court_case_id VARCHAR(50),
+    court_case_id VARCHAR(250),
     tag_id INT,
     PRIMARY KEY (court_case_id, tag_id),
     FOREIGN KEY (court_case_id) REFERENCES court_case(court_case_id),
@@ -80,7 +80,7 @@ CREATE TABLE tag_assignment (
 );
 
 CREATE TABLE judge_assignment (
-    court_case_id VARCHAR(50),
+    court_case_id VARCHAR(250),
     judge_id INT,
     PRIMARY KEY (court_case_id, judge_id),
     FOREIGN KEY (court_case_id) REFERENCES court_case(court_case_id),
@@ -88,7 +88,7 @@ CREATE TABLE judge_assignment (
 );
 
 CREATE TABLE participant_assignment (
-    court_case_id VARCHAR(50),
+    court_case_id VARCHAR(250),
     participant_id INT,
     lawyer_id INT,
     is_defendant BOOLEAN,
@@ -110,4 +110,3 @@ VALUES ('United Kingdom Supreme Court'), ('Privy Council'), ('Court of Appeal (C
 INSERT INTO verdict(verdict)
 VALUES ('Guilty'), ('Not Guilty'), ('Dismissed'), ('Acquitted'), ('Hung Jury'), ('Claimant Wins'), ('Defendant Wins'),
 ('Settlement'), ('Struck Out'), ('Appeal Dismissed'), ('Appeal Allowed'), ('Other');
-
