--- database/variables.tf
@@ -1,27 +1,0 @@
-variable "AWS_ACCESS_KEY" {
-  type = string
-}
-
-variable "AWS_SECRET_KEY" {
-  type = string
-}
-
-variable "AWS_REGION" {
-  type = string
-}
-
-variable "DB_NAME" {
-  type = string
-}
-
-variable "DB_PASSWORD" {
-  type = string
-}
-
-variable "DB_PORT" {
-  type = string
-}
-
-variable "DB_USER" {
-  type = string
-}
